+++ ui/tests/common/audit_harness.rs
@@ -7,8 +7,8 @@ use reelsynth_ui::{
     audit_all_elements, audit_center, audit_compose_panels, audit_header_clusters,
     audit_osc_sidebar_stacks, audit_panel_utilization, audit_rail_panels, audit_shell,
     compute_center_regions, draw_shell, embed_piano_in_center, record_region, AuditId,
-    ShellConfig, ShellLayout, ShellLayoutOptions, ShellMidiDevices, ShellMode, UiState,
-    APP_HEIGHT_FULL, APP_MIN_WIDTH, SPACE_SM,
+    ShellAppSettings, ShellConfig, ShellLayout, ShellLayoutOptions, ShellMidiDevices, ShellMode,
+    UiState, APP_HEIGHT_FULL, APP_MIN_WIDTH, SPACE_SM,
 };
 
 pub struct ShellAuditScenario {
@@ -18,6 +18,7 @@ pub struct ShellAuditScenario {
     pub preview: Patch,
     pub midi_names: Vec<String>,
     pub midi_selected: usize,
+    pub app_settings: Option<ShellAppSettings>,
 }
 
 impl Default for ShellAuditScenario {
@@ -34,6 +35,7 @@ impl Default for ShellAuditScenario {
             preview: Patch::factory_lead(),
             midi_names: vec!["None".to_string(), "Virtual MIDI".to_string()],
             midi_selected: 0,
+            app_settings: Some(ShellAppSettings::default()),
         }
     }
 }
@@ -120,6 +122,7 @@ pub fn run_shell_audit(scenario: ShellAuditScenario) -> ShellAuditRun {
                         &config,
                         None,
                         None,
+                        scenario.app_settings.as_mut(),
                     );
                 });
             },